backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
diffusers==0.24.0
accelerate==0.24.1
safetensors==0.4.1
controlnet-aux==0.0.7
email-validator==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
